+++ src/letters/model/letter-scope.enum.ts
@@ -1,4 +1,4 @@
 export enum LetterScope {
-  GLOBAL = 'GLOBAL',
+  PUBLIC = 'PUBLIC',
   ROOM = 'ROOM',
 }
